Refactoring & Splitting classes to different files
